requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
requests
plaid-python
scikit-learn
pandas
numpy
certifi
